% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Sangers.R
\name{Analyze_Sangers}
\alias{Analyze_Sangers}
\title{Preprocessing and annotation of Sanger sequencing of the 16S rRNA gene}
\usage{
Analyze_Sangers(
  inputfile = NULL,
  inputfolder = NULL,
  fwd_suffix = "f",
  rev_suffix = "r",
  fastas_only = FALSE,
  resfolder = "Results",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{inputfile}{LGC text file containing forward and reversed reads}

\item{inputfolder}{folder containing the ab1 files with the raw spectrograms
from an ABI capillary electrophoresis instrument. only used when raw=TRUE}

\item{fwd_suffix}{string containing the suffix of the forward reads
(defaults to f)}

\item{rev_suffix}{string containing the suffix of the reverse reads
(defaults to r)}

\item{fastas_only}{if you do not want merging to occur but want separate
fasta files to be spitted out for downstream processing in e.g. BioEdit. (
defaults to FALSE)}

\item{resfolder}{name of the folder to put the results in (defaults to
"Results/")}

\item{verbose}{print out progress (defaults to FALSE)}

\item{...}{additional arguments passed on to BioStrings::readDNAStringSet}
}
\description{
Preprocessing and annotation of Sanger sequencing of the 16S rRNA gene
}
\examples{
## File-based workflow (LGC-preprocessed)

fastademoset <- system.file("extdata","sanger_demodata.txt",
                              package = "CMETNGS",mustWork = TRUE)
fastalnres   <- Analyze_Sangers(inputfile=fastademoset,fwd_suffix="F",rev_suffix="R")

}
