% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindPad.R
\name{cbindPad}
\alias{cbindPad}
\title{function to combine unequal-length data frames (or vectors) into a data.frame}
\usage{
cbindPad(..., pad = NA)
}
\arguments{
\item{...}{vectors or data frames of unequal length}

\item{pad}{what to pad the resulting data frame with (defaults to NA)}
}
\description{
This function combines unequal-length data frames (or vectors) into a data.frame
It is copied from https://stackoverflow.com/questions/14899306/transform-a-splitted-data-frame-in-a-new-data-frame
}
\examples{
## Short example

# Load precomputed example data
#TODO: find test object

}
