% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circledraw.R
\name{circleFun}
\alias{circleFun}
\title{function to draw a circle}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{vector of length 2 with coordinates for the center (dfaults to c(0,0))}

\item{diameter}{numeric indicating diamter (defaults to 1)}

\item{npoints}{number of points (defaults to 100)}
}
\description{
This function draws a circle with given parameters
}
\examples{
## Short example

# Load precomputed example data
#TODO: find test object

}
