% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcurveg.R
\name{colcurveg}
\alias{colcurveg}
\title{Function to process mothur generated alpha-diversity collector curves}
\usage{
colcurveg(fnpref, groups, fnpost, errorbars = FALSE, printing = FALSE)
}
\arguments{
\item{fnpref}{file prefix string for the collector curve data}

\item{groups}{(ordered) character vector with names of the samples}

\item{fnpost}{name of calculator used (character)}

\item{errorbars}{logical indicating whether or not error bars should be plotted
at this stage, the code does NOT check whether or not the data for the error
bars is available so it is the users responsibility to make sure it is.}

\item{printing}{logical to indicate whether or not the plot should be displayed}
}
\value{
The underlying dataframe for the plot and the plot (if printing is set to TRUE)
}
\description{
This (legacy) function allows users to visualize mothur-generated collector
curve data using R.
}
\examples{
#TODO
}
\keyword{collector}
\keyword{curves}
\keyword{mothur,}
