% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldiss.R
\name{coldiss}
\alias{coldiss}
\title{Color plots of a dissimilarity matrix, without and with ordering}
\usage{
coldiss(D, nc = 4, byrank = TRUE, diag = FALSE)
}
\arguments{
\item{D}{dissimilarity matrix. If D is not a dissimilarity matrix (max(D)>1),
then D is divided by max(D)}

\item{nc}{number of colors (classes, defaults to 4)}

\item{byrank}{specifiy either equal-sized classes (TRUE) or equal-length
intervals (FALSE). Defaults to TRUE}

\item{diag}{print object labels on the diagonal (defaults to FALSE)}
}
\description{
This function generates color plots of dissimilarity matrices, with our
without ordering. It was supplied as supplementary data to "Numerical
ecology with R" Boccard & Legendre, 2012. And authored by Francois Gillet.
}
\examples{
## Short example

# coldiss(spe.dj, nc=9, byrank=F, diag=T)
# Original author: Francois Gillet, 23 aug 2012

}
\keyword{heatmap}
