% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2fasta.R
\name{df2fasta}
\alias{df2fasta}
\title{Conversion of data frames with correct layout to fasta files.}
\usage{
df2fasta(
  dataframe,
  filename = "fastadf",
  seqIDcolumn = "SeqIDs",
  seqColumn = "readseq"
)
}
\arguments{
\item{dataframe}{a dataframe containing sequence identifiers and sequences
in separate columns. with column names}

\item{filename}{desirec filename (defaults to fastadf)}

\item{seqIDcolumn}{string indicating the name of the column with the
sequence identifiers (defaults to SeqIDs).}

\item{seqColumn}{string indicating the name of the column with the
sequences themselves (defaults to readseq)}
}
\description{
Typically used in conjunction with fasta2dataframe.
}
\examples{
## Short example

# Load precomputed example data
data(otureps)
# write out the associated fasta
df2fasta(otureps,filename="oturep.fasta",seqIDcolumn="OTU")

}
