% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2dataframe.R
\name{fasta2dataframe}
\alias{fasta2dataframe}
\title{Conversion of (mothur) fasta files to a dataframe}
\usage{
fasta2dataframe(fastaFile, mothurrepseqs = TRUE, striparg = "Otu")
}
\arguments{
\item{fastaFile}{standard format DNA (not RNA/protein) fasta file}

\item{mothurrepseqs}{boolean indicating wether or not the file contains mothur-formatted headers for each sequence as generated by get.oturep. Defaults to TRUE.}

\item{striparg}{string to be kept as headers in more complex headers (allows for regex), ignored in case mothurrepseqs=TRUE (default)}
}
\description{
Conversion of (mothur) fasta files to a dataframe
}
\examples{
## Short example

# Load precomputed example data
#TODO: add export option

}
