% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getconsensus.R
\name{getconsensus}
\alias{getconsensus}
\title{Parsing function for consensus as found by sangeranalyseR::merge.reads}
\usage{
getconsensus(x)
}
\arguments{
\item{x}{sangeranalyseR::merged.read class object}
}
\description{
Parsing function for consensus as found by sangeranalyseR::merge.reads
}
\examples{
## Simple example

}
