% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{function to generate ggplot-like color pallettes}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
function to generate ggplot-like color pallettes
}
\examples{
## Short example

# Load precomputed example data
#TODO: find test object

}
