% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highqualgraphR.R
\name{highqualgraphR}
\alias{highqualgraphR}
\title{Publication-grade export of ggplot2 object}
\usage{
highqualgraphR(x, filename, res = 1200, pointsize = 12,
  embed = FALSE, extension = "pdf", graphicstype = "cairo",
  colorspace = "cmyk")
}
\arguments{
\item{x}{a ggplot2 object.}

\item{filename}{the filename without an extension.}

\item{res}{resolution (ppi) of the graph, defaults to 1200}

\item{pointsize}{font size (pixels), defaults to 12}

\item{embed}{a logical indicating wether in the font should be embedded
in a pdf or eps document. This is done to enhance compatibility with older
viewers or when using a non-standard font (see ?embedFonts for details). It
requires ghostscript to be installed.}

\item{extension}{character vector including: "pdf", "png", "tiff", "svg"
or "postscript"}

\item{graphicstype}{passed on to the "type" parameter in bitmap graphics
devices (defaults to "cairo", but requires cairo graphics)}

\item{colorspace}{sets the colorspace. For printing such as in papers
the default colorspace should be CMYK but for on-screen viewing srgb may
be more desirable. Can also be set to gray to enable greyscale printing.
Defaults to CMYK.}
}
\description{
This function takes a ggplot2 object and a series of extensions as input
to export it to. A high quality graph is generated complying to PLOS ONE
author guidelines.
}
\examples{
## Short example

# Load precomputed example data
#TODO: find test object

}
