% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect2.R
\name{intersect2}
\alias{intersect2}
\title{Intersection function for two character vectors}
\usage{
intersect2(x, y)
}
\arguments{
\item{x}{character vector x}

\item{y}{character vector y}
}
\description{
Intersection function for two character vectors
}
\examples{
## Simple example

}
