% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KimPalettes.R
\name{KimPalettes}
\alias{KimPalettes}
\title{Request discrete color palettes created by Kim De Paepe}
\usage{
KimPalettes(n)
}
\arguments{
\item{n}{number of colors in the plot (needs to be in the pre-defined list
in the description)}
}
\value{
a character vector with hexadecimal color definitions of the desired
 length
}
\description{
Function to select a discrete color palette as created by Kim De Paepe using
to create color palettes: http://tools.medialab.sciences-po.fr/iwanthue/ .
The options are 2, 3, 4,5, 6,7,8,9,10,11,12,15,16,20,30 and 100-color palettes.
}
\examples{
## Short example

fourcolours <- KimPalettes(4)
plot(runif(4),cex=3,pch=16,col=fourcolours)
}
