% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makebargraphrawggplot2.R
\name{makebargraphrawggplot2}
\alias{makebargraphrawggplot2}
\title{ggplot2-based cummulative stacked bargraphs for plotting taxonomy}
\usage{
makebargraphrawggplot2(tax, shared, topn = 8, taxlevel = c("Regnum",
  "Phylum", "Classis", "Ordo", "Familia", "Genus", "Species"),
  shared.abs = T, tax.prob = F, tresh = 85, samples,
  viridfill = FALSE, plot = TRUE, ...)
}
\arguments{
\item{tax}{is a taxonomy file, raw.}

\item{shared}{is a shared (OTU-by-sample) file, filtered per sample group.}

\item{topn}{is the top number of taxa selected to display (maximum 11,
default 8)}

\item{taxlevel}{is either one of  "Regnum", "Phylum", "Classis",
"Ordo",  "Familia", "Genus"}

\item{shared.abs}{is a logical showing wether the shared file already
consists out of direct Illumina counts, or is really the raw csv from
Ramiro (default TRUE)}

\item{tax.prob}{shows whether the taxonomy file contains probabilites
(default FALSE)}

\item{tresh}{is a threshold for the probability taxonomy duplicate matching
(to be implemented, default 85)}

\item{samples}{is an optional string for the title of the generated bargraph
(e.g. "MOB","OLAND","Fecal samples", defualt empty)}

\item{viridfill}{whether or not a discrete viridis fill pallette should be
used (defaults to FALSE)}

\item{plot}{states wether or not to print the plot within the function
(default TRUE)}

\item{...}{optional arguments passed on to ggplot call}
}
\value{
a list of 3 objects: ggplotdf (the underlying data for the ggplot),
 othertax (an abundance-sorted dataframe of the OTUs summarized in "other"),
 ggplotobj (the generated ggplot2 object)
}
\description{
This function uses a taxonomy file (i.e. an RDP naive bayesian fixed-rank
asignment detail) and a shared file to create 100% summed stacked bargraphs.
Only the top-n OTUs (in relative abundance) among all samples are displayed.
All other OTUs are summed into a part of the graph designated as "other".
The resulting object contains information about the composition of "other".
}
\examples{
## Short example

# Load precomputed example data
#TODO: find testset

}
\keyword{diversity}
\keyword{taxonomic}
