% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeEDABargraphs.R
\name{MakeEDABargraphs}
\alias{MakeEDABargraphs}
\title{Function to automatically generate a few Phyloseq bargraphs on the data}
\usage{
MakeEDABargraphs(shared, taxonomy, retfphyloseq = TRUE)
}
\arguments{
\item{shared}{file location string for the shared file
(preferably absolute path)}

\item{taxonomy}{file location string for the taxonomy file (preferably
absolute path)}

\item{retfphyloseq}{wheter or not to return a phyloseq object (defaults
to true)}
}
\value{
A phyloseq object and three plots in the current directory
}
\description{
This function uses a shared file (i.e. a mothur-formatted OTU contingency
table) and a taxonomy file (i.e. an RDP naive bayesian fixed-rank
asignment detail) to generate three stacked barcharts in the folder it is
executed in. The function returns a phyloseq object as well by default.
}
\examples{

shareddataset <- system.file("extdata","large_shared_file.shared",
                              package = "CMETNGS",mustWork = TRUE)
taxonomydataset <- system.file("extdata","large_OTU_basedtax.taxonomy",
                              package = "CMETNGS",mustWork = TRUE)
edagraphs <- MakeEDABargraphs(shareddataset,taxonomydataset)


}
\keyword{EDA}
\keyword{bargraph,}
\keyword{mothur}
\keyword{stacked}
