% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myround.R
\name{myround}
\alias{myround}
\title{Alternative rounding function}
\usage{
myround(x)
}
\arguments{
\item{x}{numeric vector to be rounded}
}
\description{
Alternative rounding function
}
\examples{
## Short example

# Load precomputed example data
#TODO: find test object

}
