% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordin_annotatedplot.R
\name{ordin_annotatedplot}
\alias{ordin_annotatedplot}
\title{Create an annotated ordination graph}
\usage{
ordin_annotatedplot(ordinobj, metadata, location = "bottomleft",
  addLabels = TRUE, addLegend = TRUE)
}
\arguments{
\item{ordinobj}{a vegan-created ordination object (i.e. ordinobj$points must
exist)}

\item{metadata}{a valid CMET metadata object, containing minimally the sample
names, Factor1 and Factor2.}

\item{location}{where to put the legend on the plot, defaults to "bottomleft"}

\item{addLabels}{wether or not to add sample labels to the ordination (
defaults to TRUE)}

\item{addLegend}{a logical, indicating wether or not to add a legend to the
plot. Defaults to TRUE.}
}
\value{
A list with the colvector and pchvector
}
\description{
This function is a wrapper around CMETNGS::ordin_annotate and CMETNGS::ordin_legend:
it uses a vegan ordination object (e.g. from metaMDS) and a valid metadata
table with Factor1 and Factor2 to create an annotated plot.
}
\examples{
## Short example

# Load precomputed example data
library(vegan)
data(dune)
dune.dis <- vegdist(wisconsin(dune))
dune.mds <- cmdscale(dune.dis, eig = TRUE)
#create completely made-up metadata for dune
dune.md <- data.frame(Factor1=c(rep("Dune1",nrow(dune)/2),
rep("Dune2",nrow(dune)/2)),Factor2=c(rep("Grass",7),rep("Shrub",7),
rep("Tree",6)))
ordin_annotatedplot(dune.mds,dune.md)
}
\keyword{ordination}
