% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordin_legend.R
\name{ordin_legend}
\alias{ordin_legend}
\title{Create legend for base vegan ordination plots}
\usage{
ordin_legend(metadata, location = "bottomleft")
}
\arguments{
\item{metadata}{a valid CMET metadata object, containing minimally the sample
names, Factor1 and Factor2.}

\item{location}{where to put the legend on the plot, defaults to "bottomleft"}
}
\value{
A list with the colvector and pchvector
}
\description{
This function uses a vegan ordination object (e.g. from metaMDS) and
a valid metadata table with Factor1 and Factor2 and is meant to be used in
conjunction with CMETNGS::ordin_annotate.
}
\examples{
## Short example

# Load precomputed example data
library(vegan)
data(dune)
dune.dis <- vegdist(wisconsin(dune))
dune.mds <- cmdscale(dune.dis, eig = TRUE)
#create completely made-up metadata for dune
dune.md <- data.frame(Factor1=c(rep("Dune1",nrow(dune)/2),
rep("Dune2",nrow(dune)/2)),Factor2=c(rep("Grass",7),rep("Shrub",7),
rep("Tree",6)))
veclist <- ordin_annotate(dune.mds,dune.md)
dune.mds$species <- wascores(dune.mds$points, dune, expand = TRUE)
pl <- ordiplot(dune.mds, type = "none")
points(pl, "sites", pch=veclist$pchvect, col=veclist$colvect, bg="yellow")
ordipointlabel(dune.mds,display = "sites",add=TRUE)
ordin_legend(dune.md,location="topright")
}
\keyword{ordination}
