% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preformattax.R
\name{preformattax}
\alias{preformattax}
\title{reformatter of taxonomy files from mothur for use in R}
\usage{
preformattax(taxonomy)
}
\arguments{
\item{taxonomy}{is a taxonomy file, raw.}
}
\description{
This function uses a taxonomy file (i.e. an RDP naive bayesian fixed-rank
asignment detail) and pre-processes it into a data frame for use within R.
}
\examples{
## Short example

# Load precomputed example data
#TODO: find testset

}
\keyword{converter}
