% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmerger.R
\name{readmerger}
\alias{readmerger}
\title{Support function for merging abifldr objects primary sequences}
\usage{
readmerger(
  fwdfile,
  revfile,
  samplename,
  onefile = FALSE,
  outputfolder = "Results"
)
}
\arguments{
\item{fwdfile}{path to forward ab1 file}

\item{revfile}{path to reverse ab1 file}

\item{samplename}{name for fasta identifier/file name for merged read}

\item{onefile}{whether or not to write out one fasta file with all consensus
reads (defaults to FALSE)}

\item{outputfolder}{path to output folder}
}
\value{
a list containing the sample, consensus length and consensus seq
}
\description{
Support function for merging abifldr objects primary sequences
}
\examples{
##Simple example

}
