% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Sangers.R
\name{Analyze_Sangers}
\alias{Analyze_Sangers}
\title{Preprocessing and annotation of Sanger sequencing of the 16S rRNA gene}
\usage{
Analyze_Sangers(
  inputfile = NULL,
  inputfolder = NULL,
  fwd_suffix = "F",
  rev_suffix = "R",
  fastas_only = FALSE,
  resfolder = "Results",
  verbose = FALSE,
  trim.cutoff = 0.005,
  ncores = 2,
  mintrim = 300L,
  ...
)
}
\arguments{
\item{inputfile}{LGC text file containing forward and reversed reads}

\item{inputfolder}{folder containing the ab1 files with the raw spectrograms
from an ABI capillary electrophoresis instrument. only used when raw=TRUE}

\item{fwd_suffix}{string containing the suffix of the forward reads
(defaults to F)}

\item{rev_suffix}{string containing the suffix of the reverse reads
(defaults to R)}

\item{fastas_only}{if you do not want merging to occur but want separate
fasta files to be spitted out for downstream processing in e.g. BioEdit. (
defaults to FALSE)}

\item{resfolder}{name of the folder to put the results in (defaults to
"Results/")}

\item{verbose}{print out progress (defaults to FALSE)}

\item{trim.cutoff}{sangeranalyseR::summarise.abi.folder argument (defaults to 5e-3)}

\item{ncores}{number of cores for sangeranalyseR::summarise.abi.folder (
defaults to 2, and will be set to 1 on windows)}

\item{mintrim}{minimal length of trimmed reads in the folder workflow to
retain before merging (defaults to 300)}

\item{...}{additional arguments passed on to BioStrings::readDNAStringSet or
sangeranalyseR::summarise.abi.folder in case inputfile==NULL and inputfolder
is specified}
}
\value{
either a list of the consensus reads, forward and reverse reads or a summarised abi folder
}
\description{
Preprocessing and annotation of Sanger sequencing of the 16S rRNA gene
}
\examples{
##File-based workflow (LGC-preprocessed)

##Folder-based workflow

# TODO: find proper example set
}
