% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preformattax.R
\name{preformattax}
\alias{preformattax}
\title{Reformatter of taxonomy files from mothur for use in R}
\usage{
preformattax(taxonomy, OTUtax = TRUE, probab = TRUE, keepProbab = TRUE,
  keepOTU = FALSE, keepSize = FALSE)
}
\arguments{
\item{taxonomy}{is a taxonomy file, raw.}

\item{OTUtax}{is a logical indicating wether we are dealing with a
classify.otu (TRUE) or classify.seqs (FALSE) taxonomy file. Defaults to
TRUE.}

\item{probab}{is a logical indicating whether the classify command was run
with probs =T (i.e. probabilities or bootstrap confidences are included).
Defaults to TRUE (= probs are in the taxonomy file). Currently ignored for
if FALSE.}

\item{keepProbab}{is a logical indicating whether or not to keep the columns
with the probabilities. Defaults to TRUE. Ignored when probab is FALSE}

\item{keepOTU}{is a logical indicating whether or not to include a column with
the OTU names called OTU. Defaults to FALSE.}

\item{keepSize}{is a logical indicating whether or not to include a column
that contains the total read count within the OTU. Defaults to FALSE.}
}
\value{
A dataframe with a splitted taxonomy file.
The Size column is dropped.
}
\description{
This function uses a taxonomy file (i.e. an RDP naive bayesian fixed-rank
asignment detail) and pre-processes it into a data frame for use within R.
}
\examples{
## Short example

# Load precomputed example data
#TODO: find testset

}
\keyword{converter}
